fix(nav): move routers to utils folder
