+++ lib/presentation/utils/routes/app_router.dart
@@ -3,12 +3,12 @@ import 'package:flutter/foundation.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:vasture/presentation/providers/tutorial_provider.dart';
+import 'package:vasture/presentation/components/loading/loading_overlay.dart';
 import 'package:vasture/presentation/providers/user_provider.dart';
-import 'package:vasture/presentation/screens/camera/camera_screen.dart';
+import 'package:vasture/presentation/screens/camera/camera_page_view_screen.dart';
 import 'package:vasture/presentation/screens/introduction/introduction_screen.dart';
 
-import 'package:vasture/routes/routes.dart';
+import 'package:vasture/presentation/utils/routes/routes.dart';
 import 'package:vasture/presentation/components/navigation_bar/app_navigation_bar.dart';
 import 'package:vasture/presentation/screens/weather_book/weather_book_screen.dart';
 import 'package:vasture/presentation/screens/weather/weather_screen.dart';
@@ -16,6 +16,7 @@ import 'package:vasture/presentation/screens/weather/weather_screen.dart';
 part 'branch/weather_branch.dart';
 part 'branch/weather_book_branch.dart';
 part 'routers/introduction_router.dart';
+part 'routers/camera_router.dart';
 part 'app_router.g.dart';
 
 final rootNavigatorKey = GlobalKey<NavigatorState>();
@@ -28,15 +29,24 @@ GoRouter goRouter(Ref ref) {
     routes: $appRoutes,
     debugLogDiagnostics: kDebugMode,
     redirect: (context, state) async {
-      final tutorialState = ref.read(tutorialProvider);
-      final userState = ref.read(userProvider);
+      final userId = await ref.read(sharedPrefsDataSourceProvider).getUserIdFromRouter();
+      final isTutorialCompleted =
+          await ref.read(sharedPrefsDataSourceProvider).isTutorialCompleted();
 
       // NOTE チュートリアルが終わっていて、ユーザーが存在するとWeatherScreenへ！
-      if (tutorialState.isCompleted && userState.user != null) {
-        return AppRoutes.weather;
-      } else {
+      if (isTutorialCompleted && userId.isNotEmpty) {
+        return null;
+      }
+
+      if ((isTutorialCompleted && userId.isEmpty) && state.matchedLocation == AppRoutes.weather) {
         return AppRoutes.introduction;
       }
+
+      if (!isTutorialCompleted && state.matchedLocation == AppRoutes.weather) {
+        return AppRoutes.introduction;
+      }
+
+      return null;
     },
   );
 }
@@ -48,7 +58,6 @@ GoRouter goRouter(Ref ref) {
     weatherBookStatefulShellBranch,
   ],
 )
-
 class MainShellRouteData extends StatefulShellRouteData {
   const MainShellRouteData();
 
@@ -58,6 +67,8 @@ class MainShellRouteData extends StatefulShellRouteData {
     GoRouterState state,
     StatefulNavigationShell navigationShell,
   ) {
-    return AppNavigationBar(navigationShell: navigationShell);
+    return LoadingOverlay(
+      child: AppNavigationBar(navigationShell: navigationShell),
+    );
   }
 }
