+++ lib/presentation/utils/routes/routes.dart
@@ -1,7 +1,9 @@
+// NOTE: ナビゲーションの定義
+
 class AppRoutes {
   static const weather = '/weather';
   static const introduction = '/introduction';
-  static const camera = '/camera';
+  static const cameraPageView = '/cameraPageView';
   static const confirmation = '/confirmation';
   static const matchedSky = '/matchedSky';
   static const weatherBook = '/weatherBook';
@@ -9,3 +11,7 @@ class AppRoutes {
   static const photoDetail = '/photoDetail';
   static const setting = '/setting';
 }
+
+// class WeatherAppRoutes {
+//   static const cameraPageView = 'cameraPageView';
+// }
